pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mili-ui"
version = "1.0.5"
authors = [{ name = "damusss" }]
description = "Minimal immediate-mode python user interface library"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ['pygame-ce >= 2.5.2']
license = { file = "LICENSE.txt" }
keywords = ['gui', 'ui', 'pygame', 'pygame-ce']

[project.urls]
Homepage = "https://github.com/damusss/mili"
Issues = "https://github.com/damusss/mili/issues"
Documentation = "https://github.com/damusss/mili/blob/main/guide/guide.md"

[tool.setuptools]
packages = ['mili', 'mili._utils']
